fun_projects/web/musical_time_machine/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
spotipy>=2.23.0
python-dotenv>=1.0.1